requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
plotly>=5.14.1
matplotlib>=3.7.1
numpy>=1.24.3
sqlalchemy>=2.0.15
pymysql>=1.0.3
cryptography>=40.0.2
python-dotenv>=1.0.0
pillow>=9.5.0
streamlit-authenticator>=0.2.2
streamlit-option-menu>=0.3.2
streamlit-calendar>=0.1.0
bcrypt>=4.0.1
pdfplumber>=0.9.0
openpyxl>=3.1.2
python-dateutil>=2.8.2
scikit-learn>=1.2.2
faker>=18.9.0
plotly-calplot>=0.1.16
jsonpickle>=3.0.1